create question in exam management

Closes #285

See merge request !192

+++ ipho_exam/views.py
@@ -44,7 +44,7 @@ from ipho_exam import qml, tex, pdf, iphocode, qquery, fonts, cached_responses, 
 from ipho_exam.response import render_odt_response
 from ipho_print import printer
 
-from ipho_exam.forms import LanguageForm, FigureForm, TranslationForm, PDFNodeForm, FeedbackForm, AdminBlockForm, AdminBlockAttributeFormSet, AdminBlockAttributeHelper, SubmissionAssignForm, AssignTranslationForm, TranslationImportForm, AdminImportForm, PrintDocsForm, ScanForm, ExtraSheetForm
+from ipho_exam.forms import LanguageForm, FigureForm, TranslationForm, ExamQuestionForm, DeleteForm, VersionNodeForm, PDFNodeForm, FeedbackForm, AdminBlockForm, AdminBlockAttributeFormSet, AdminBlockAttributeHelper, SubmissionAssignForm, AssignTranslationForm, TranslationImportForm, AdminImportForm, PrintDocsForm, ScanForm, ExtraSheetForm
 
 import ipho_exam
 from ipho_exam import tasks
@@ -743,6 +743,102 @@ def figure_export(request, fig_id, output_format='svg', lang_id=None):
         Figure.to_pdf(fig_svg, tmpfile)
         return HttpResponse(open(tmpfile), content_type="application/pdf")
 
+@permission_required('ipho_core.is_staff')
+def admin_add_question(request, exam_id):
+    if not request.is_ajax:
+        raise Exception('TODO: implement small template page for handling without Ajax.')
+
+    ## Question section
+    question_form = ExamQuestionForm(request.POST or None)
+    exam = get_object_or_404(Exam, id=exam_id)
+
+    if question_form.is_valid():
+        question_form.instance.exam = exam
+        if question_form.cleaned_data.get('type') != Question.ANSWER:
+            question_form.instance.working_pages = 0
+        question_form.save()
+
+        return JsonResponse({
+                    'type'    : 'submit',
+                    'success' : True,
+                    'message' : '<strong>Question created!</strong> The new question '+question_form.cleaned_data.get('name')+' has successfully been created.',
+                    'exam_id': exam_id,
+                })
+
+    form_html = render_crispy_form(question_form)
+    return JsonResponse({
+                'title'   : 'Add new quesiton',
+                'form'    : form_html,
+                'submit'  : 'Create',
+                'success' : False,
+            })
+
+@permission_required('ipho_core.is_staff')
+def admin_delete_question(request, exam_id, question_id):
+    if not request.is_ajax:
+        raise Exception('TODO: implement small template page for handling without Ajax.')
+    
+    delete_form = DeleteForm(request.POST or None)
+
+    delete_message = 'This action <strong>CANNOT</strong> be undone. <strong>All versions and all translations</strong> of this question will be lost.'
+
+    if delete_form.is_valid():
+        question = get_object_or_404(Question, id=question_id)
+        if question.name == delete_form.cleaned_data.get('verify'):
+            question.delete()
+            return JsonResponse({
+                        'type'    : 'submit',
+                        'success' : True,
+                        'message' : '<strong>Question deleted!</strong> The question '+question.name+' has been deleted.',
+                        'exam_id': exam_id,
+                    })
+        else:
+            form_html = render_crispy_form(delete_form)
+            return JsonResponse({
+                        'title'   : 'Are you sure?',
+                        'form'    : '<div class="alert alert-warning alert-dismissible"><button type="button" class="close" data-dismiss="alert"><span>&times;</span></button>'\
+                                    +'<strong>Try again!</strong> The question name does not match.</div>'\
+                                    +delete_message\
+                                    +form_html,
+                        'success' : False,
+                    })
+
+    form_html = render_crispy_form(delete_form)
+    return JsonResponse({
+                'title'   : 'Are you sure?',
+                'form'    : delete_message\
+                            +form_html,
+                'success' : False,
+            })
+
+@permission_required('ipho_core.is_staff')
+def admin_edit_question(request, exam_id, question_id):
+    if not request.is_ajax:
+        raise Exception('TODO: implement small template page for handling without Ajax.')
+
+    ## Question section
+    instance = get_object_or_404(Question, pk=question_id)
+    question_form = ExamQuestionForm(request.POST or None, instance=instance)
+
+    if question_form.is_valid():
+        if question_form.cleaned_data.get('type') != Question.ANSWER:
+            question_form.instance.working_pages = 0
+        question = question_form.save()
+
+        return JsonResponse({
+                    'type'    : 'submit',
+                    'success' : True,
+                    'message' : '<strong>Question modified!</strong> The question '+question.name+' has successfully been modified.',
+                    'exam_id': exam_id,
+                })
+
+    form_html = render_crispy_form(question_form)
+    return JsonResponse({
+                'title'   : 'Edit quesiton',
+                'form'    : form_html,
+                'submit'  : 'Save',
+                'success' : False,
+            })
 
 @permission_required('ipho_core.is_staff')
 @ensure_csrf_cookie
@@ -833,8 +929,37 @@ def admin_delete_version(request, exam_id, question_id, version_num):
     else:
         raise Exception('Only versioned node can be deleted')
 
-    node.delete()
-    return HttpResponseRedirect(reverse('exam:admin'))
+    delete_form = DeleteForm(request.POST or None)
+
+    delete_message = 'This action <strong>CANNOT</strong> be undone. All changes made in this version will be lost.'
+
+    if delete_form.is_valid():
+        if question.name == delete_form.cleaned_data.get('verify'):
+            node.delete()
+            return JsonResponse({
+                        'type'    : 'submit',
+                        'success' : True,
+                        'message' : '<strong>Version deleted!</strong> Version v'+str(node.version)+' of question '+question.name+' has been deleted.',
+                        'exam_id': exam_id,
+                    })
+        else:
+            form_html = render_crispy_form(delete_form)
+            return JsonResponse({
+                        'title'   : 'Are you sure?',
+                        'form'    : '<div class="alert alert-warning alert-dismissible"><button type="button" class="close" data-dismiss="alert"><span>&times;</span></button>'\
+                                    +'<strong>Try again!</strong> The question name does not match.</div>'\
+                                    +delete_message\
+                                    +form_html,
+                        'success' : False,
+                    })
+
+    form_html = render_crispy_form(delete_form)
+    return JsonResponse({
+                'title'   : 'Are you sure?',
+                'form'    : delete_message\
+                            +form_html,
+                'success' : False,
+            })
 
 @permission_required('ipho_core.is_staff')
 def admin_accept_version(request, exam_id, question_id, version_num, compare_version=None):
@@ -844,6 +969,12 @@ def admin_accept_version(request, exam_id, question_id, version_num, compare_ver
     question = get_object_or_404(Question, id=question_id)
 
     lang = get_object_or_404(Language, id=lang_id)
+
+    if not VersionNode.objects.filter(question=question, language=lang, status__in=['S','C']):
+        node = get_object_or_404(VersionNode, question=question, language=lang, status='P', version=version_num)
+        node.status = 'S'
+        node.save()
+        return HttpResponseRedirect(reverse('exam:admin'))
 
     if compare_version is None:
         compare_node = VersionNode.objects.filter(question=question, language=lang, status__in=['S','C']).order_by('-version')[0]
@@ -906,6 +1037,38 @@ def admin_publish_version(request, exam_id, question_id, version_num):
     node.status = 'C'
     node.save()
     return HttpResponseRedirect(reverse('exam:admin'))
+
+@permission_required('ipho_core.is_staff')
+def admin_settag_version(request, exam_id, question_id, version_num):
+    if not request.is_ajax:
+        raise Exception('TODO: implement small template page for handling without Ajax.')
+
+    lang_id = OFFICIAL_LANGUAGE
+
+    exam = get_object_or_404(Exam, id=exam_id)
+    question = get_object_or_404(Question, id=question_id)
+    lang = get_object_or_404(Language, id=lang_id)
+
+    ## Version section
+    node = get_object_or_404(VersionNode, question=question, language=lang, version=version_num)
+    node_form = VersionNodeForm(request.POST or None, instance=node)
+
+    if node_form.is_valid():
+        node_form.save()
+        return JsonResponse({
+                    'type'    : 'submit',
+                    'success' : True,
+                    'message' : '<strong>Tag set!</strong> The tag has successfully been modified.',
+                    'exam_id': exam_id,
+                })
+
+    form_html = render_crispy_form(node_form)
+    return JsonResponse({
+                'title'   : 'Set version tag',
+                'form'    : form_html,
+                'submit'  : 'Save',
+                'success' : False,
+            })
 
 @permission_required('ipho_core.is_staff')
 @ensure_csrf_cookie
